Delete all the catalogue-related code we no longer need
