+++ common/stacks/src/main/scala/weco/api/stacks/models/StacksUserIdentifier.scala
@@ -1,0 +1,3 @@
+package weco.api.stacks.models
+
+case class StacksUserIdentifier(value: String)
